Revert "Cruz nicolas rama final"
