+++ tp2_Grupo22/pom.xml
@@ -37,11 +37,6 @@
 			<artifactId>spring-boot-starter-test</artifactId>
 			<scope>test</scope>
 		</dependency>
-		<dependency>
-			<groupId>ar.edu.unju.fi</groupId>
-			<artifactId>pedidos</artifactId>
-			<version>0.0.1-SNAPSHOT</version>
-		</dependency>
 	</dependencies>
 
 	<build>
