+++ tp2_Grupo22/src/main/java/ar/edu/unju/fi/controller/ServicioPaseoController.java
@@ -1,4 +1,4 @@
-  package ar.edu.unju.fi.controller;
+package ar.edu.unju.fi.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -12,4 +12,5 @@ import org.springframework.web.bind.annotation.GetMapping;
 			 
 		}
 	}
- 
+
+
