+++ tp2_Grupo22/src/main/java/ar/edu/unju/fi/model/Sucursal.java
@@ -1,4 +1,4 @@
- package ar.edu.unju.fi.model;
+package ar.edu.unju.fi.model;
 
 import java.time.LocalDate;
