+++ tp2_Grupo22/src/main/resources/templates/layouts/header.html
@@ -13,10 +13,11 @@
 				<li class="imain-nav-item"><a th:href="@{/productos}">PRODUCTOS</a> </li>
 				<li class="imain-nav-item"><a th:href="@{/consejos}">CONSEJOS PARA LA SALUD</a> </li>
 				<li class="imain-nav-item"><a th:href="@{/paseo}">SERVICIOS DE PASEOS</a> </li>
-				<li class="imain-nav-item"><a th:href="@{/sucursal}">SUCURSALES</a> </li>
+				<li class="imain-nav-item"><a th:href="@{/sucursales}">SUCURSALES</a> </li>
 				<li class="imain-nav-item"><a th:href="@{/contactos}">CONTACTANOS</a> </li>
+
 			</ul>
 		</nav>
 	</header>
 </body>
-</html> 
+</html>
