+++ tp2_Grupo22/src/main/resources/templates/serviciospaseo.html
@@ -67,4 +67,4 @@
 		
 	<footer th:replace="~{layouts/footer :: footer}"></footer>
 </body>
-</html>   
+</html>
