+++ tp2_Grupo22/src/main/resources/templates/sucursales.html
@@ -1,5 +1,5 @@
- <!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org" lang="es">
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org" lang="es" <!-- class="html_sucursales" -->>
 
 <head>
 	
@@ -13,13 +13,13 @@
 		
 		<header th:replace="~{layouts/header :: header}"></header>	
 	
-			<section id="content-sucursal">
+			<section id="content-sucursal"
 			
 				<a th:href="@{/sucursal/nuevo}">Nueva Sucursal</a><br>
 						<!-- TABLA DE TURNOS DE PASEADORES -->
-					<table border="1" class="tabla_sucursal">
+					<table border="1" calss="tabla_sucursal">
 								
-								<caption id="cap">Visita Nuestras Sucursales</caption>
+								<caption id="cap"><h2> Visita Nuestras Sucursales </h2></caption>
 								
 								<tr>
 									<th class="th">Tienda</th>
@@ -40,10 +40,39 @@
 									<td th:text="${sucu.teledono}"></td>
 									<td th:text="${sucu.grupo}"></td>
 								</tr>
-							
+								<!--<tr>
+									<td class="td">Mora PetShop</td>
+								 	<td class="td"><img class="image" th:src="@{images\Cruz-image\tienda1.jpg}" alt="Fachada de la primera tienda" title="foto tienda principal"></td>
+								 	<td class="td"><iframe  src="https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d3639.2330439098!2d-65.29737921114504!3d-24.1986164!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x941b0f2afccbd6fd%3A0x1ab3d331e73b57f8!2sMora%20Pet%20Shop!5e0!3m2!1ses-419!2sar!4v1682273704021!5m2!1ses-419!2sar" width="600" height="450" style="border:0;" allowfullscreen="" loading="lazy" referrerpolicy="no-referrer-when-downgrade"></iframe></td>
+									<td class="td"><a>10/05/2021</a></td>
+									<td class="td"><a href="mailto:mora@petshop.com">mora@petshop.com</a></td>
+									<td class="td"><a>+543885896541</a></td>
+									<td class="td"><a href="https://chat.whatsapp.com/BlrmT1OHgIs2cBkuEkgMCW"><img class="image" th:src="@{images\Cruz-image\wpp.png}"</a></td>
+								</tr>
+								 	
+								<tr>
+									<td class="td">Mundo PetShop</td>
+								 	<td class="td"><img class="image" th:src="@{images\Cruz-image\tienda2.jpg}" alt="Fachada de la segunda tienda" title="foto segunda tienda"></td>
+								 	<td class="td"><iframe src="https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d14557.614906065855!2d-65.30120818261719!3d-24.192635999999986!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x941b0f57489d9543%3A0x318d1ece83f1aedc!2sMundo%20Maskota%60s!5e0!3m2!1ses-419!2sar!4v1682272740759!5m2!1ses-419!2sar" width="600" height="450" style="border:0;" allowfullscreen="" loading="lazy" referrerpolicy="no-referrer-when-downgrade"></iframe></td>
+									<td class="td"><a>10/05/2021</a></td>
+									<td class="td"><a href="mailto:mundo@petshop.com">mundo@petshop.com</a></td>
+									<td class="td"><a>+543885321456</a></td>
+									<td class="td"><a href="https://chat.whatsapp.com/BlrmT1OHgIs2cBkuEkgMCW"><img class="image" th:src="@{images\Cruz-image\wpp.png}"</a></td>
+								</tr>
+								
+								<tr>
+									<td class="td">DonCacho PetShop</td>
+									<td class="td"><img class="image" th:src="@{images\Cruz-image\tienda3.jpg}" alt="Fachada de la tercera tienda" title="foto tercera tienda"></td>	
+									<td class="td"><iframe src="https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d3639.316675315633!2d-65.29985502496969!3d-24.19568628520185!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x941b0f94383b867d%3A0xb5791169187694e5!2sTienda%20Para%20Mascotas!5e0!3m2!1ses-419!2sar!4v1682273357672!5m2!1ses-419!2sar" width="600" height="450" style="border:0;" allowfullscreen="" loading="lazy" referrerpolicy="no-referrer-when-downgrade"></iframe></td>
+									<td class="td"><a>10/05/2021</a></td>
+									<td class="td"><a href="mailto:doncacho@petshop.com">doncacho@petshop.com</a></td>
+									<td class="td"><a>+543885963852</a></td>
+									<td class="td"><a href="https://chat.whatsapp.com/BlrmT1OHgIs2cBkuEkgMCW"><img class="image" th:src="@{images\Cruz-image\wpp.png}"</a></td>
+								</tr>	
+								-->
 					</table>
 			</section>
 		<footer th:replace="~{layouts/footer :: footer}"></footer>
 	
 </body>
-</html> 
+</html>
